Support QXL hotplug_mode_update
Summary:
QXL is a driver used mainly in VMs.
This driver, it creates a new preferred mode upon resize instead of just
offering the one and resizing it. This makes Plasma not resize when it should
as described on the bug report below.
This patch allows kscreen to know if the driver considers we should be
following the preferred mode.

https://bugzilla.redhat.com/show_bug.cgi?id=1290586

Test Plan: Tested extensively on one of these vms and outside.

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D15789