Do hash and hex name based output hashes
Summary:
The hash of an output is based on its name in case no EDID information is
available. In order to have the same hash format as when EDID is available
do the same hash operation to the name as to the EDID checksum.

Introduce the new  `hashMd5` function and deprecate the previous `hash`
function. This way we can use the new function, which does hash the
output names correctly, in cases where it makes sense and no user configs
break.

Test Plan: Compiles.

Reviewers: #plasma, lbeltrame

Subscribers: lbeltrame, mart, plasma-devel

Tags: #plasma

Maniphest Tasks: T10028

Differential Revision: https://phabricator.kde.org/D17007