Wayland: output device name as create function argument only
Summary:
With saving the pointers it is not necessary anymore to save the Wayland name
in WaylandOutput except when creating the output device. Simplify the code by
removing the member variable and providing the name as an argument to the
create output device function.

Test Plan: Manually in Wayland session.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23025