Wayland: set KWayland outputs data in output class
Summary:
When sending a new config to KWayland set the output specific data of the
configuration in WaylandOutput instead of WaylandConfig. This simplifies
the code and keeps responsibilities better separated.

Test Plan:
Changing various parameters in Wayland session. Mode setting is half broken
like it was before on my test setup.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23028