RandR: Replicate outputs with XRender extension
Summary:
This provides means to set replicas and queries them on X11 by transforming
crtcs with the XRender extension.

There is a heuristic at play to detect possible replications and currently
the aspect ratio is not perserved but the image stretched. Using different
values for the transformation matrix also the former should be possible.

Test Plan: Manually together with patch to KScreen.

Reviewers: #kwin, #plasma

Subscribers: davidedmundson, plasma-devel

Tags: #plasma

Maniphest Tasks: T11222

Differential Revision: https://phabricator.kde.org/D23663