Add enforced mode size getter
Summary:
Introduce little helper to API, which gives back either the current mode size
or if no current mode is set the preferred mode or if there is no preferred
mode as last resort the first one in the map if the map is not empty.

In a small behavior change `geometry()` then can still give a valid geometry
back even if no current mode is set at the moment.

Test Plan: Compiles.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24037