docs: add contributing guide
Summary:
This adds a document providing information on how to contribute to libkscreen
and additionally a minimal readme file pointing to it.

The contributing guide is the analogous to a newly added one in KScreen.

It includes clarifications about the used coding style and introduces a Commit
Message Guideline based on the KDE Commit Policy and the Conventional Commits
specification.

Reviewers: #plasma

Subscribers: davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25064