feat: add auto rotate and tablet mode information
Summary:
This adds feature flags if a backend supports auto rotations. While this is not
directly a function of a backend on X11 auto rotation is not tested fully yet
and there are problems with touch input. This way a client can query if auto
rotation is fully functional on a backend.

For the KWayland backend auto rotation and tablet mode information querying is
already wired up but for now deactivated until all KWin patches have been
landed.

Test Plan:
Compiles and tests with an XPS 2-in-1, which sadly does not send events about
the tablet state.

Reviewers: #plasma, apol

Reviewed By: apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26281