feat: replace replication source with logical size API
Summary:
Just overriding the logical size of an output that replicates another one is
simpler than trying to send a relation between both objects to the display
server and in case of X11 it is not possible.

Wires up support for that in the X11 backend.

Test Plan: Compiles. Wayland replication tested.

Reviewers: #kwin

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26309