fix(randr): make sure crtc exists
When an output gets enabled we directly want to set the logical size. But at
this point because of the order of calls in XRandRConfig the crtc is not yet
set on the output.

To fix this in a quick way just provide the free crtc as an argument. Long-term
change the order of calls.