fix: transfer tablet mode properties
Summary:
I forgot to copy over the tablet-mode-available and -engaged data from one
config to the next one. Because of that the data was not available to the
daemon.

Also serialize the tablet mode properties in the Serializer. The Serializer
is not used on Wayland afaik but if at one point for example the RandR backend
gains auto rotation support we need to have it in the Serializer.

BUG: 418739

Test Plan: Manually with hard-coded tablet mode availability in KWin.

Reviewers: #plasma, bshah

Reviewed By: #plasma, bshah

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28106