wayland: Fix output hotplugging
When an output is hotplugged, kscreen kded module will find a matching
config and check whether it can be applied by calling
Config::canBeApplied().

The Config::canBeApplied() function will check whether the current output
setup satisfies various properties specified in the config, e.g. modes,
etc. One of those things is the maximum number of active outputs -
Screen::maxActiveOutputsCount().

When an output is added, the backend needs to call
Screen::setMaxActiveOutputCount() to ensure that the kded module will be
able to properly determine whether a config candidate is applicable.

With output device v2 changes, one of the setters got lost, this change
brings it back.