fix(randr): screen is dirty when switching display modes
Because of the order of calls in XRandRConfig, screen is dirty when switching display modes.

This patch removes the intermediateScreenSize.
Swap the change order of crtc and screen size, and disable crtc before that.

The previous method was to first set an intermediate screen size then set the crtc and finally set the target screen size.
The modified method is to disable crtc first, then change the screen size and finally set crtc.
This change can effectively solve the problem of blurry screen when switching display modes.