Introduce KScreenDpms
This is a new library that only has one class abstract the control of
the system's DPMS.

This code was extracted from PowerDevil which was already abstracting
this functionality. Here we are doing in a generic way so it can be used
elsewhere.

This also ports the kscreen-doctor dpms code to use it instead of its
wayland-only application, so now it will also work on x11.