Ban KScreen::Output::setId(): Remove method from API
Output ID is the essence of Output object's identity. It is used as a
key in OutputList (which is actually a QMap). It must not change,
otherwise funny things may happen. This patch series removes all usages
of setId one by one, and then drops the method entirely from API.