backends/kwayland: Remove unbalanced unblockSignals()
This unblockSignals() is not balanced with any blockSignals(). We could
add a blockSignals() but it will make things more fragile, for example
what if the connection times out. Blocking signals should be unnecessary
because output device properties are sent in batches followed by done events
and the WaylandConfig avoids emitting the configChanged signal while
some outputs are still initializing.

This fixes powerdevil crashing when kwin restarts.