Make AbstractBackend::setConfig() return a QFuture
Screen configuration uses nested event loops in a few places and they
are a source of crashes. With QPAs such as QtWayland, it's even preferred
to avoid nested event loops as they can break screen initialization.

Draft: not tested and I'm not sure about a few API details, but it compiles so it should work