Allow for async in-process backends
The public facing API of kscreen allows for async operations.
For DBus-based out of process backends, this is a call to another
process.

Both in-process and out-of-process backends share the same abstraction
to apply the configuration. This is fine for X11 where it's out of
process, but causes the wayland implementation to spawn a nested event
loop. Nested event loops are bad.

This change introduces a path for backends to provide an async reponse
to set config operation, and map that to the existing async public
interface.

The wayland backend uses an additional QObject to scope the signals to
cleanup and auto-disconnect when either signal fires.

BUG: 483072