backends/kwayland: drop the output order protocol for a new "priority" event
When a screen is mirrored, KWin needs to remove it from the output order for
plasmashell, however, that doesn't mean its position in the output order list
should be forgotten.
To make this work, the output order protocol is now only used for plasmashell,
and the priority of an output device is directly communicated in the output
device protocol instead.
This also has the advantage of dropping some code here.